// src/app/Models/documents.ts
export class Documents {
    id: string | undefined; 
    customer: string | undefined; 
    invoiceNumber: string | undefined; 
    type: string | undefined; 
    invoiceOwner: string | undefined; 
    caseNumber: string | undefined; 
    income: number | undefined; 
    timeConsuming: number | undefined; 
    drive: number | undefined; 
    month: number | undefined; 
    invoiceStatus: string | undefined; 
    createDate: Date | undefined; 
    settlementDate: Date | undefined; 
    modifyDate: Date |undefined; 
    newInvoice: boolean |undefined; 
    preAccept: boolean |undefined;
    accepted: boolean |undefined;
}